bar-chart and change code to api
